dbapi: extract common code from the 2 Cursors
Noticed while going to implement executemany, that
the two cursors share a whole lot of similar code i.e.
* autocommit_off_cursor.py
* autocommit_on_cursor.py

so this change refactors and extracts common code
from both cursors, bringing it into base_cursor.BaseCursor
which all Cursors will now inherit from.
