+++ spanner/dbapi/autocommit_on_cursor.py
@@ -6,6 +6,7 @@
 
 import google.api_core.exceptions as grpc_exceptions
 
+from .base_cursor import BaseCursor
 from .exceptions import (
     Error, IntegrityError, OperationalError, ProgrammingError,
 )
@@ -19,45 +20,17 @@ from .utils import PeekIterator
 _UNSET_COUNT = -1
 
 
-class Cursor(object):
-    def __init__(self, db_handle=None):
-        self.__itr = None
-        self.__res = None
-        self.__row_count = _UNSET_COUNT
-        self.__db_handle = db_handle
-        self.__last_op = None
-        self.__closed = False
-
-        # arraysize is a readable and writable property mandated
-        # by PEP-0249 https://www.python.org/dev/peps/pep-0249/#arraysize
-        # It determines the results of .fetchmany
-        self.arraysize = 1
-
-    @property
-    def description(self):
-        if not (self.__res and self.__res.metadata):
-            return None
-
-        row_type = self.__res.metadata.row_type
-        columns = []
-        for field in row_type.fields:
-            columns.append(Column(name=field.name, type_code=field.type.code))
-        return tuple(columns)
-
-    @property
-    def rowcount(self):
-        return self.__row_count
-
-    def __raise_if_already_closed(self):
+class Cursor(BaseCursor):
+    def _raise_if_already_closed(self):
         """
         Raises an exception if attempting to use an already closed connection.
         """
-        if self.__closed:
+        if self._closed:
             raise Error('attempting to use an already closed connection')
 
     def close(self):
         self.__clear()
-        self.__closed = True
+        self._closed = True
 
     def execute(self, sql, args=None):
         """
@@ -69,23 +42,23 @@ class Cursor(object):
         Returns:
             None
         """
-        self.__raise_if_already_closed()
+        self._raise_if_already_closed()
 
-        if not self.__db_handle:
+        if not self._connection:
             raise ProgrammingError('Cursor is not connected to the database')
 
-        self.__res = None
+        self._res = None
 
         # Classify whether this is a read-only SQL statement.
         try:
             classification = classify_stmt(sql)
             if classification == STMT_DDL:
-                self.__db_handle.append_ddl_statement(sql)
+                self._connection.append_ddl_statement(sql)
                 return
 
             # For every other operation, we've got to ensure that
             # any prior DDL statements were run.
-            self.__run_prior_DDL_statements()
+            self._run_prior_DDL_statements()
 
             if classification == STMT_NON_UPDATING:
                 self.__handle_DQL(sql, args or None)
@@ -101,7 +74,7 @@ class Cursor(object):
             raise OperationalError(e.details if hasattr(e, 'details') else e)
 
     def __handle_update(self, sql, params):
-        self.__db_handle.in_transaction(
+        self._connection.in_transaction(
             self.__do_execute_update,
             sql, params,
         )
@@ -111,9 +84,9 @@ class Cursor(object):
         sql, params = sql_pyformat_args_to_spanner(sql, params)
 
         res = transaction.execute_update(sql, params=params, param_types=get_param_types(params))
-        self.__itr = None
+        self._itr = None
         if type(res) == int:
-            self.__row_count = res
+            self._row_count = res
 
         return res
 
@@ -136,7 +109,7 @@ class Cursor(object):
         if parts.get('homogenous'):
             # The common case of multiple values being passed in
             # non-complex pyformat args and need to be uploaded in one RPC.
-            return self.__db_handle.in_transaction(
+            return self._connection.in_transaction(
                 self.__do_execute_insert_homogenous,
                 parts,
             )
@@ -144,7 +117,7 @@ class Cursor(object):
             # All the other cases that are esoteric and need
             #   transaction.execute_sql
             sql_params_list = parts.get('sql_params_list')
-            return self.__db_handle.in_transaction(
+            return self._connection.in_transaction(
                 self.__do_execute_insert_heterogenous,
                 sql_params_list,
             )
@@ -166,14 +139,14 @@ class Cursor(object):
         return transaction.insert(table, columns, values)
 
     def __handle_DQL(self, sql, params):
-        with self.__db_handle.read_snapshot() as snapshot:
+        with self._connection.read_snapshot() as snapshot:
             # Reference
             #  https://googleapis.dev/python/spanner/latest/session-api.html#google.cloud.spanner_v1.session.Session.execute_sql
             sql, params = sql_pyformat_args_to_spanner(sql, params)
             res = snapshot.execute_sql(sql, params=params, param_types=get_param_types(params))
             if type(res) == int:
-                self.__row_count = res
-                self.__itr = None
+                self._row_count = res
+                self._itr = None
             else:
                 # Immediately using:
                 #   iter(response)
@@ -183,13 +156,13 @@ class Cursor(object):
                 # are for .fetchone() with those that would result in
                 # many items returns a RuntimeError if .fetchone() is
                 # invoked and vice versa.
-                self.__res = res
+                self._res = res
                 # Read the first element so that StreamedResult can
                 # return the metadata after a DQL statement. See issue #155.
-                self.__itr = PeekIterator(self.__res)
+                self._itr = PeekIterator(self._res)
                 # Unfortunately, Spanner doesn't seem to send back
                 # information about the number of rows available.
-                self.__row_count = _UNSET_COUNT
+                self._row_count = _UNSET_COUNT
 
     def __enter__(self):
         return self
@@ -198,125 +171,12 @@ class Cursor(object):
         self.__clear()
 
     def __clear(self):
-        self.__db_handle = None
+        self._connection = None
 
     def executemany(self, operation, seq_of_params):
-        self.__raise_if_already_closed()
+        self._raise_if_already_closed()
 
-        if not self.__db_handle:
+        if not self._connection:
             raise ProgrammingError('Cursor is not connected to the database')
 
         raise ProgrammingError('Unimplemented')
-
-    def __next__(self):
-        if self.__itr is None:
-            raise ProgrammingError('no results to return')
-        return next(self.__itr)
-
-    def __iter__(self):
-        if self.__itr is None:
-            raise ProgrammingError('no results to return')
-        return self.__itr
-
-    def fetchone(self):
-        self.__raise_if_already_closed()
-
-        try:
-            return next(self)
-        except StopIteration:
-            return None
-
-    def fetchall(self):
-        self.__raise_if_already_closed()
-
-        return list(self.__iter__())
-
-    def fetchmany(self, size=None):
-        """
-        Fetch the next set of rows of a query result, returning a sequence of sequences.
-        An empty sequence is returned when no more rows are available.
-        Args:
-            size: optional integer to determine the maximum number of results to fetch.
-        Raises:
-            Error if the previous call to .execute*() did not produce any result set
-            or if no call was issued yet.
-        """
-        self.__raise_if_already_closed()
-
-        if size is None:
-            size = self.arraysize
-
-        items = []
-        for i in range(size):
-            try:
-                items.append(tuple(self.__next__()))
-            except StopIteration:
-                break
-
-        return items
-
-    @property
-    def lastrowid(self):
-        return None
-
-    def setinputsizes(sizes):
-        raise ProgrammingError('Unimplemented')
-
-    def setoutputsize(size, column=None):
-        raise ProgrammingError('Unimplemented')
-
-    def __run_prior_DDL_statements(self):
-        return self.__db_handle.run_prior_DDL_statements()
-
-    def list_tables(self):
-        return self.__db_handle.list_tables()
-
-    def run_sql_in_snapshot(self, sql):
-        return self.__db_handle.run_sql_in_snapshot(sql)
-
-    def get_table_column_schema(self, table_name):
-        return self.__db_handle.get_table_column_schema(table_name)
-
-
-class Column:
-    def __init__(self, name, type_code, display_size=None, internal_size=None,
-                 precision=None, scale=None, null_ok=False):
-        self.name = name
-        self.type_code = type_code
-        self.display_size = display_size
-        self.internal_size = internal_size
-        self.precision = precision
-        self.scale = scale
-        self.null_ok = null_ok
-
-    def __repr__(self):
-        return self.__str__()
-
-    def __getitem__(self, index):
-        if index == 0:
-            return self.name
-        elif index == 1:
-            return self.type_code
-        elif index == 2:
-            return self.display_size
-        elif index == 3:
-            return self.internal_size
-        elif index == 4:
-            return self.precision
-        elif index == 5:
-            return self.scale
-        elif index == 6:
-            return self.null_ok
-
-    def __str__(self):
-        rstr = ', '.join([field for field in [
-            "name='%s'" % self.name,
-            "type_code=%d" % self.type_code,
-            None if not self.display_size else "display_size='%s'" % self.display_size,
-            None if not self.internal_size else "internal_size='%s'" % self.internal_size,
-            None if not self.precision else "precision='%s'" % self.precision,
-            None if not self.scale else "scale='%s'" % self.scale,
-            None if not self.null_ok else "null_ok='%s'" % self.null_ok,
-        ] if field])
-
-        return 'Column(%s)' % rstr
